Actualizar Ejercicio Division Numeros con el Do/While

+++ Ejercicio3_Guia13/DivisionNumero.java
@@ -1,9 +1,14 @@
 package Ejercicio3_Guia13;
 
+// Defina una clase llamada DivisionNumero. En el método main utilice un Scanner para leer dos
+// números en forma de cadena. A continuación, utilice el método parseInt() de la clase Integer,
+// para convertir las cadenas al tipo int y guardarlas en dos variables de tipo int. Por ultimo realizar
+// una división con los dos numeros y mostrar el resultado.
+
 public class DivisionNumero {
     public void division(int numA,int numB){
         
-        int  resultado = numA / numB ;
+        double  resultado = numA / numB ;
         System.out.println("La division de los numeros es: " + resultado);
         
     }
